add finish solution
