+++ src/types/FooterType.ts
@@ -1,0 +1,11 @@
+import { Dispatch, SetStateAction } from 'react';
+import { FilterType } from './FilterType';
+import { Todo } from './Todo';
+
+export type FooterType = {
+  todos: Todo[];
+  setTodos: Dispatch<SetStateAction<Todo[]>>;
+  setCurrentFilter: (filter: FilterType) => void;
+  setError: Dispatch<SetStateAction<string>>;
+  setIsLoadingIds: Dispatch<SetStateAction<number[]>>;
+};
