+++ src/types/NotificationProps.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export type NotificationProps = {
+  error: string;
+  setError: Dispatch<SetStateAction<string>>;
+};
